requirements.txt
opencv-python~=4.8.1.78
paho-mqtt~=1.6.1
ultralytics~=8.0.227
